Changed chat style, added button for make a new conversations from profile page

+++ src/main/java/web/servlet/changers/ChangerMessageServlet.java
@@ -1,6 +1,8 @@
 package web.servlet.changers;
 
 import com.google.inject.Inject;
+import com.google.inject.Singleton;
+import model.Message;
 import model.User;
 import service.ConversationService;
 
@@ -12,6 +14,7 @@ import java.io.IOException;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 
+@Singleton
 public class ChangerMessageServlet extends HttpServlet {
 
     private static final org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(ChangerMessageServlet.class);
@@ -36,13 +39,31 @@ public class ChangerMessageServlet extends HttpServlet {
         String action = req.getParameter("action");
         String conv_idS = req.getParameter("conv_id");
         String msg_idS = req.getParameter("msg_id");
+        String otherUserIDS = req.getParameter("otherUserID");
+        String new_msg = req.getParameter("new_msg");
         switch (action) {
+            case "createConversation": {
+                if (otherUserIDS != null && !otherUserIDS.isEmpty()) {
+                    long otherUserID = Long.parseLong(otherUserIDS);
+                    Message new_message = Message.builder()
+                            .sender_id(currentUserID)
+                            .ts_action(Timestamp.valueOf(LocalDateTime.now()))
+                            .message(new_msg)
+                            .build();
+                    conversationService.createConversation(currentUserID,otherUserID,new_message);
+                    return;
+                } else {
+                    resp.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+                    req.getRequestDispatcher("/WEB-INF/not_found.jsp").forward(req, resp);
+                    return;
+                }
+            }
             case "deleteConversation": {
-                if (conv_idS != null) {
-                    long conv_id = Long.parseLong(req.getParameter("conv_id"));
+                if (conv_idS != null && !conv_idS.isEmpty()) {
+                    long conv_id = Long.parseLong(conv_idS);
                     conversationService.deleteConversation(conv_id);
                     resp.setStatus(HttpServletResponse.SC_OK);
-                    resp.sendRedirect(req.getContextPath() + "/message");
+                    resp.sendRedirect(req.getContextPath() + "/messages");
                     return;
                 } else {
                     resp.setStatus(HttpServletResponse.SC_BAD_REQUEST);
@@ -51,12 +72,12 @@ public class ChangerMessageServlet extends HttpServlet {
                 }
             }
             case "deleteMessage": {
-                if (conv_idS !=null && msg_idS != null) {
+                if (conv_idS != null && !conv_idS.isEmpty() && msg_idS != null && !msg_idS.isEmpty()) {
                     long conv_id = Long.parseLong(conv_idS);
                     long msg_id = Long.parseLong(msg_idS);
                     conversationService.deleteMessage(msg_id);
                     resp.setStatus(HttpServletResponse.SC_OK);
-                    resp.sendRedirect(req.getContextPath() + "/message_" + conv_idS);
+                    resp.sendRedirect(req.getContextPath() + "/messages_" + conv_idS);
                     return;
                 } else {
                     resp.setStatus(HttpServletResponse.SC_BAD_REQUEST);
